+++ src/components/layout/PageContent/PageContent.tsx
@@ -1,7 +1,9 @@
 import type { FC } from "react";
+import React from "react";
 import classNames from "classnames";
 import classes from "./PageContent.module.scss";
 import type { PageContentProps } from "./types";
+import { AppErrorBoundary } from "@/components/layout/AppErrorBoundary";
 
 const PageContent: FC<PageContentProps> = ({
   children,
@@ -16,7 +18,7 @@ const PageContent: FC<PageContentProps> = ({
           [classes.alignLeft]: container === "medium" && align === "left",
         })}
       >
-        {children}
+        <AppErrorBoundary>{children}</AppErrorBoundary>
       </div>
     </div>
   );