+++ src/features/instances/components/AccessGroupFilter/AccessGroupFilter.tsx
@@ -2,6 +2,7 @@ import type { FC } from "react";
 import type { GroupedOption } from "@/components/filter";
 import { TableFilter } from "@/components/filter";
 import usePageParams from "@/hooks/usePageParams";
+import useSetDynamicFilterValidation from "@/hooks/useDynamicFilterValidation";
 
 interface AccessGroupFilterProps {
   readonly options: GroupedOption[];
@@ -9,6 +10,11 @@ interface AccessGroupFilterProps {
 
 const AccessGroupFilter: FC<AccessGroupFilterProps> = ({ options }) => {
   const { accessGroups, setPageParams } = usePageParams();
+
+  useSetDynamicFilterValidation(
+    "accessGroups",
+    options.map((opt) => opt.value),
+  );
 
   return (
     <TableFilter