+++ src/features/package-profiles/components/PackageProfileHeader/PackageProfileHeader.tsx
@@ -1,15 +1,14 @@
+import { TableFilterChips } from "@/components/filter";
 import HeaderWithSearch from "@/components/form/HeaderWithSearch";
-import usePageParams from "@/hooks/usePageParams";
 import type { FC } from "react";
 
 const PackageProfileHeader: FC = () => {
-  const { setPageParams } = usePageParams();
-
-  const handleSearch = (searchText: string) => {
-    setPageParams({ search: searchText });
-  };
-
-  return <HeaderWithSearch onSearch={handleSearch} />;
+  return (
+    <>
+      <HeaderWithSearch />
+      <TableFilterChips filtersToDisplay={["search"]} />
+    </>
+  );
 };
 
 export default PackageProfileHeader;