+++ src/features/repository-profiles/components/RepositoryProfileHeader/RepositoryProfileHeader.tsx
@@ -1,0 +1,14 @@
+import { TableFilterChips } from "@/components/filter";
+import HeaderWithSearch from "@/components/form/HeaderWithSearch";
+import type { FC } from "react";
+
+const RepositoryProfileHeader: FC = () => {
+  return (
+    <>
+      <HeaderWithSearch />
+      <TableFilterChips filtersToDisplay={["search"]} />
+    </>
+  );
+};
+
+export default RepositoryProfileHeader;