+++ src/main.tsx
@@ -8,19 +8,27 @@ import NotifyProvider from "@/context/notify";
 import ReactQueryProvider from "@/context/reactQuery";
 import App from "./App";
 import { ROOT_PATH } from "@/constants";
+import * as Sentry from "@sentry/browser";
+import AppErrorBoundary from "./components/layout/AppErrorBoundary/AppErrorBoundary";
+
+Sentry.init({
+  dsn: "https://55a60b44ddfd4ca5a94a8a3bac2d5052@sentry.is.canonical.com//85",
+});
 
 createRoot(document.getElementById("root") as HTMLElement).render(
   <StrictMode>
-    <BrowserRouter basename={ROOT_PATH}>
-      <EnvProvider>
-        <NotifyProvider>
-          <ReactQueryProvider>
-            <AuthProvider>
-              <App />
-            </AuthProvider>
-          </ReactQueryProvider>
-        </NotifyProvider>
-      </EnvProvider>
-    </BrowserRouter>
+    <AppErrorBoundary>
+      <BrowserRouter basename={ROOT_PATH}>
+        <EnvProvider>
+          <NotifyProvider>
+            <ReactQueryProvider>
+              <AuthProvider>
+                <App />
+              </AuthProvider>
+            </ReactQueryProvider>
+          </NotifyProvider>
+        </EnvProvider>
+      </BrowserRouter>
+    </AppErrorBoundary>
   </StrictMode>,
 );